+++ cloud-service/src/repo/template.rs
@@ -16,10 +16,10 @@ pub struct TemplateRecord {
     pub template_id: Uuid,
     pub name: String,
     pub size: i32,
-    pub version: i32,
+    pub version: i64,
     pub user_template: String,
     pub protected_template: String,
-    pub protector_version: Option<i32>,
+    pub protector_version: Option<i64>,
     pub metadata: String,
     pub project_id: Uuid,
 }
@@ -32,7 +32,7 @@ impl TryFrom<TemplateRecord> for crate::model::Template {
             .map_err(|e| format!("Invalid Template Metadata: {}", e))?;
         let versioned_template_id: VersionedTemplateId = VersionedTemplateId {
             template_id: TemplateId(value.template_id),
-            version: value.version,
+            version: value.version as u64,
         };
         let protected_template_id: ProtectedTemplateId = ProtectedTemplateId {
             versioned_template_id: versioned_template_id.clone(),
@@ -58,7 +58,7 @@ impl From<crate::model::Template> for TemplateRecord {
             template_id: value.versioned_template_id.template_id.0,
             name: value.template_name.0,
             size: value.template_size,
-            version: value.versioned_template_id.version,
+            version: value.versioned_template_id.version as i64,
             user_template: value.versioned_template_id.slug(),
             protected_template: value.protected_template_id.slug(),
             protector_version: None,
@@ -82,7 +82,7 @@ pub trait TemplateRepo {
     async fn get_by_version(
         &self,
         template_id: &Uuid,
-        version: i32,
+        version: u64,
     ) -> Result<Option<TemplateRecord>, RepoError>;
 
     async fn get_by_project(&self, project_id: &Uuid) -> Result<Vec<TemplateRecord>, RepoError>;
@@ -168,13 +168,13 @@ impl TemplateRepo for DbTemplateRepo<sqlx::Postgres> {
     async fn get_by_version(
         &self,
         template_id: &Uuid,
-        version: i32,
+        version: u64,
     ) -> Result<Option<TemplateRecord>, RepoError> {
         sqlx::query_as::<_, TemplateRecord>(
             "SELECT template_id, version, project_id, name, size, user_template, protected_template, protector_version, jsonb_pretty(templates.metadata) AS metadata  FROM templates WHERE template_id = $1 AND version = $2",
         )
         .bind(template_id)
-        .bind(version)
+        .bind(version as i64)
         .fetch_optional(self.db_pool.deref())
         .await
         .map_err(|e| e.into())
@@ -328,13 +328,13 @@ impl TemplateRepo for DbTemplateRepo<sqlx::Sqlite> {
     async fn get_by_version(
         &self,
         template_id: &Uuid,
-        version: i32,
+        version: u64,
     ) -> Result<Option<TemplateRecord>, RepoError> {
         sqlx::query_as::<_, TemplateRecord>(
             "SELECT template_id, version, project_id, name, size, user_template, protected_template, protector_version, jsonb_pretty(templates.metadata) AS metadata  FROM templates WHERE template_id = $1 AND version = $2",
         )
             .bind(template_id)
-            .bind(version)
+            .bind(version as i64)
             .fetch_optional(self.db_pool.deref())
             .await
             .map_err(|e| e.into())
